changed unit test framework from unittest to pytest
