+++ requirements.txt
@@ -1,0 +1,6 @@
+coverage==7.8.0
+iniconfig==2.1.0
+packaging==25.0
+pluggy==1.6.0
+pytest==8.3.5
+pytest-cov==6.1.1
